Move config into lib/config.rs
