+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod app;
 pub mod args;
 pub mod cli;
+pub mod config;
 pub mod execution;
 pub mod telemetry;
